Rename test to df_cleanup
Sync to renamed function
